Conflicts:
	Druzi/views.py

+++ Druzi/views.py
@@ -3,7 +3,7 @@ from django.db.models import Count
 import operator
 import re
 from datetime import datetime
-from Druzi.models import Activity, Enrollment, Tag, TagAppear
+from Druzi.models import Activity, Enrollment, Tag, TagAppear, Rating
 from django.contrib import messages
 from django.core import serializers
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
@@ -12,6 +12,9 @@ from django.http import HttpResponseRedirect, HttpResponse
 from django.shortcuts import render
 from .forms import ActivityForm
 from django.contrib.auth.decorators import login_required
+from django.core import serializers
+from django.utils import timezone
+
 
 # Create your views here.
 
@@ -46,7 +49,7 @@ def activity_creation(request):
             activity.description = description
             activity.save()
             messages.success(request, 'Se ha creado correctamente la actividad')
-            return HttpResponseRedirect('/')
+            return HttpResponseRedirect(reverse('activity_details', kwargs={'slug' : activity.get_slug , 'id': activity.id}))
 
     # if a GET (or any other method) we'll create a blank form
     else:
@@ -69,7 +72,7 @@ def activity_pagination(request, page="1"):
         list = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": list, "page": page, "last": paginator.num_pages, 'url': 'activity_list_page'})
+                  {"activity_list": list, 'url': 'activity_list_page'})
 
 
 def activity_ultimos_propuestos_pagination(request, page="1"):
@@ -86,8 +89,7 @@ def activity_ultimos_propuestos_pagination(request, page="1"):
         list = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": list, "page": int(page), "last": paginator.num_pages,
-                   'url': 'activity_ultimos_propuestos_pagination'})
+                  {"activity_list": list, 'url': 'activity_ultimos_propuestos_pagination'})
 
 
 def activity_mas_buscados_pagination(request, page="1"):
@@ -104,8 +106,7 @@ def activity_mas_buscados_pagination(request, page="1"):
         list = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": list, "page": int(page), "last": paginator.num_pages,
-                   'url': 'activity_mas_buscados_pagination'})
+                  {"activity_list": list, 'url': 'activity_mas_buscados_pagination'})
 
 
 def activity_mas_baratos_pagination(request, page="1"):
@@ -122,8 +123,7 @@ def activity_mas_baratos_pagination(request, page="1"):
         list = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": list, "page": int(page), "last": paginator.num_pages,
-                   'url': 'activity_mas_baratos_pagination'})
+                  {"activity_list": list, 'url': 'activity_mas_baratos_pagination'})
 
 
 def activity_mas_propuestos_pagination(request, page="1"):
@@ -141,13 +141,12 @@ def activity_mas_propuestos_pagination(request, page="1"):
         lista = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": lista, "page": int(page), "last": paginator.num_pages,
-                   'url': 'activity_mas_propuestos_pagination'})
+                  {"activity_list": lista, 'url': 'activity_mas_propuestos_pagination'})
 
 
 @login_required
 def activity_mylist_pagination(request, page="1"):
-    activity_list = Activity.objects.filter(participants=request.user).order_by('activity_date')
+    activity_list = Activity.objects.filter(participants=request.user).order_by('-activity_date')
     paginator = Paginator(activity_list, 10)
     try:
         lista = paginator.page(page)
@@ -159,12 +158,11 @@ def activity_mylist_pagination(request, page="1"):
         lista = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": lista, "page": int(page), "last": paginator.num_pages,
-                   'url': 'activity_mylist_pagination'})
+                  {"activity_list": lista, 'url': 'activity_mylist_pagination'})
 
 
 @login_required
-def activity_enrrolment(request, id):
+def activity_enrrolment(request, slug, id):
     activity = Activity.objects.get(id=id)
     enrollment = Enrollment(activity=activity, user=request.user)
     count_participants = activity.participants.all().count()
@@ -181,15 +179,19 @@ def activity_enrrolment(request, id):
             return HttpResponseRedirect(reverse('activity_details', kwargs={'id': id}))
 
 @login_required
-def activity_unenrrolment(request, id):
+def activity_unenrrolment(request, slug, id):
     activity = Activity.objects.get(id=id)
     enrollment = Enrollment.objects.get(activity=activity, user=request.user)
-    enrollment.delete()
-    messages.success(request, "Te has borrado correctamente a la actividad")
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    if activity.activity_date <= timezone.now() :
+        messages.warning(request, "No te puedes borrar, la actividad ya esta cerrada")
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    else :
+        enrollment.delete()
+        messages.success(request, "Te has borrado correctamente a la actividad")
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
 
 @login_required
-def activity_repeat(request, id):
+def activity_repeat(request, slug, id):
     # if this is a POST request we need to process the form data
     if request.method == 'POST':
         # create a form instance and populate it with data from the request:
@@ -216,7 +218,7 @@ def activity_repeat(request, id):
             activity.description = description
             activity.save()
             messages.success(request, 'Se ha creado correctamente la actividad')
-            return HttpResponseRedirect('/')
+            return HttpResponseRedirect(reverse('activity_details', kwargs={'id': activity.id}))
 
     # if a GET (or any other method) we'll create a blank form
     else:
@@ -243,7 +245,7 @@ def search(request):
         first = True
         list = []
         if query["query"] == "":
-            messages.info(request, "Tienes que introducir algun valor de busqueda")
+            messages.info(request, "Tienes que introducir algun valor de busqueda, si ya lo has introducido, recuerda pulsar al ENTER antes de dar al boton Buscar")
             return HttpResponseRedirect('/')
         else:
             for i, criteria in enumerate(query["query"].split(',')):
@@ -269,14 +271,35 @@ def search_tag(request, tag):
     return render(request, 'webapp/activity_list.html', {"activity_list": list})
 
 
-def activity_details(request, id):
+def activity_details(request, slug, id):
     activity = Activity.objects.get(id=id)
     activity.visit_count = activity.visit_count + 1
     activity.save()
-    return render(request, 'webapp/activity_details.html', {"activity": activity})
+    user_rating = activity.rating_set.filter(user = request.user)
+    if user_rating.count()>0:
+        user_rating = user_rating[0]
+    else:
+        user_rating = None
+    return render(request, 'webapp/activity_details.html', {"activity": activity, "user_rating" : user_rating})
+
+def stars_post(request, id):
+    activity = Activity.objects.get(id=id)
+    rating = Rating.objects.filter(activity=activity, user=request.user)
+    if rating.count() == 0:
+        if request.method == 'POST':
+            rating_val = request.POST['rating']
+            rating_val = float(rating_val)
+            rating_obj = Rating(activity=activity, user = request.user, rating=rating_val)
+            activity.sum_rating = activity.sum_rating + rating_val
+            activity.count_rating = activity.count_rating + 1
+            activity.save()
+            rating_obj.save()
+        return HttpResponse("ok", content_type='application/text')
+    else:
+        return HttpResponse("Ya habias votado esta actividad", content_type='application/text')
 
 @login_required
-def activity_list_repeat(request, id, page="1"):
+def activity_list_repeat(request, slug, id, page="1"):
     activity_list = Activity.objects.filter(Q(parent_id=id) | Q(id=id))
     paginator = Paginator(activity_list, 10)
     try:
@@ -289,5 +312,70 @@ def activity_list_repeat(request, id, page="1"):
         lista = paginator.page(paginator.num_pages)
 
     return render(request, 'webapp/activity_list.html',
-                  {"activity_list": lista, "page": int(page), "last": paginator.num_pages,
-                   'url': 'activity_list_repeat', 'origin' : int(id)})
+                  {"activity_list": lista, 'url': 'activity_list_repeat', 'origin' : int(id)})
+                   
+@login_required
+def activity_remove(request, slug, id):
+    activity = Activity.objects.get(id=id)
+    if request.user == activity.user_own:
+        if activity.is_enable:
+            activity.delete()
+            messages.success(request, "Has borrado correctamente a la actividad")
+        else:
+            messages.warning(request, "Estas intentando borrar una actividad en un tiempo excesivo")
+    else:
+        messages.warning(request, "Estas intentando borrar una actividad que no has creado tu")
+    return HttpResponseRedirect('/')
+
+@login_required
+def activity_modify(request, slug, id):
+    activity = Activity.objects.get(id=id)
+    if request.user == activity.user_own:
+        if activity.is_enable:
+            # if this is a POST request we need to process the form data
+            if request.method == 'POST':
+                # create a form instance and populate it with data from the request:
+                form = ActivityForm(request.POST)
+                # check whether it's valid:
+                if form.is_valid():
+                    activity_form = form.save(commit=False)
+                    
+                    activity.title = activity_form.title
+                    activity.description = activity_form.description
+                    activity.activity_date = activity_form.activity_date
+                    activity.position = activity_form.position
+                    activity.place_name = activity_form.place_name
+                    activity.price = activity_form.price
+                    activity.limit_participants = activity_form.limit_participants
+                    activity.tags.clear()
+                    
+                    tags = re.compile("\S*#(?:\[[^\]]+\]|\S+)").findall(activity.description)
+                    description = activity.description
+                    for t in tags:
+                        pos = activity.description.index(t)
+                        temp = Tag.objects.get_or_create(name=t[1:])[0]
+                        description = description.replace(t, '<a href="/search/tag/' + temp.name + '/">' + t + '</a>')
+                        temp.count = temp.count + 1
+                        temp.save()
+                        if temp:
+                            TagAppear(activity=activity, tag=temp, position=pos).save()
+                        else:
+                            tag = Tag(name=t[1:]).save()
+                            TagAppear(activity=activity, tag=tag, position=pos).save()
+                    activity.description = description
+                    activity.save()
+                    messages.success(request, 'Se ha modificado correctamente la actividad')
+                    return HttpResponseRedirect(reverse('activity_details', kwargs={'slug' : slug , 'id': activity.id}))
+        
+            # if a GET (or any other method) we'll create a blank form
+            else:
+                activity = Activity.objects.get(id=id)
+                activity.description = activity.get_description_text
+                form = ActivityForm(instance=activity)
+            return render(request, 'webapp/actvity_creation.html', {'form': form})
+        else:
+            messages.warning(request, "Estas intentando modificar una actividad en un tiempo excesivo")
+    else:
+        messages.warning(request, "Estas intentando modificar una actividad que no has creado tu")
+    return HttpResponseRedirect('/')
+
